Moved initialisation to !core
All modules and servers are now initialised in !core in on_server_available and on_server_join, using the _data.py in each module when possible. All current modules now check the 'activated' flag in the data.json file for the module for the server and do not do anything if 'activated' is false.